requirements.txt
streamlit==1.26.0
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
xlsxwriter>=3.0.0
